Move package to mozilla.appservices.fxaclient
