+++ components/fxa-client/android/src/main/java/mozilla/appservices/fxaclient/Config.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fxaclient.internal
+package mozilla.appservices.fxaclient
 
 /**
  * Config represents the server endpoint configurations needed for the
