+++ components/fxa-client/android/src/main/java/mozilla/appservices/fxaclient/FirefoxAccount.kt
@@ -2,10 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fxaclient.internal
+package mozilla.appservices.fxaclient
 
 import android.util.Log
 import com.sun.jna.Pointer
+import mozilla.appservices.fxaclient.rust.FxaHandle
+import mozilla.appservices.fxaclient.rust.FxaClient
+import mozilla.appservices.fxaclient.rust.Error
 import java.util.concurrent.atomic.AtomicLong
 
 /**
